requirements.txt
Django
dj-database-url==1.2.0
django-environ==0.11.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-yasg==1.21.7